requirements.txt
pylint
psycopg2-binary
confluent-kafka
python-dotenv